requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
nltk
openai
python-dotenv